Refactor seat service imports and models to unify naming conventions and remove deprecated files
