+++ backend/src/__tests__/services/calendarClockServiceTests.ts
@@ -1,6 +1,6 @@
 // calendarClockServiceTests.ts run via "npx tsx src/__tests__/services/calendarClockServiceTests.ts"
 import SurveyService from '../../services/survey.service.js';
-import SeatService from '../../services/copilot.seats.service.js';
+import SeatService from '../../services/seats.service.js';
 import metricsService from '../../services/metrics.service.js';
 import TeamsService from '../../services/teams.service.js';
 import { runSurveysForDate } from '../__mock__/survey-gen/runSurveyGenerator.js';
