+++ backend/src/__tests__/services/seatServiceInsertStandalone.ts
@@ -1,9 +1,9 @@
 // seatServiceInsertStandalone.ts  run via "npx tsx src/__tests__/services/seatServiceInsertStandalone.ts"
-import SeatService from '../../services/copilot.seats.service.js';
+import SeatService from '../../services/seats.service.js';
 import { generateStatefulMetrics } from '../__mock__/seats-gen/runSeatsGenerator.js';
 import Database from '../../database.js';
 import 'dotenv/config';
-import { SeatType } from 'models/copilot.seats.model.js';
+import { SeatType } from 'models/seats.model.js';
 
 if (!process.env.MONGODB_URI) throw new Error('MONGODB_URI is not defined');
 const database = new Database(process.env.MONGODB_URI);
