+++ backend/src/controllers/survey.controller.ts
@@ -6,7 +6,6 @@ import app from '../index.js';
 import mongoose from 'mongoose';
 import { MemberType } from 'models/teams.model.js';
 import { memoryUsage } from 'process';
-import validator from 'validator';
 
 class SurveyController {
   async updateSurveyGitHub(req: Request, res: Response): Promise<void> {
@@ -67,11 +66,9 @@ class SurveyController {
   async createSurvey(req: Request, res: Response): Promise<void> {
     try {
       const newSurvey = req.body;
-      //const member = await this.getMemberByLogin(newSurvey.userId, newSurvey.org)|| true; //needs to be fixed :"Cannot read properties of undefined (reading 'getMemberByLogin')""
-      if (true) {
-        const survey = surveyService.createSurvey(newSurvey);
-        res.status(201).json(survey);
-      }
+      // TODO: validate the user belong to the org.
+      const survey = surveyService.createSurvey(newSurvey);
+      res.status(201).json(survey);
     } catch (error) {
       res.status(500).json((error as Error).message);
       return;
@@ -169,44 +166,6 @@ class SurveyController {
       }
     } catch (error) {
       res.status(500).json(error);
-    }
-    //create a helper function that calls this api
-    //router.get('/members/:login', teamsController.getMemberByLogin);
-  }
-
-  async getMemberByLogin(loginToFind: string, orgToMatch: string): Promise<MemberType> {
-    // Ensure both values are provided.
-    if (!loginToFind || !orgToMatch) {
-      throw new Error('Both login and org must be provided');
-    }
-
-    // Trim the login input and then validate.
-    const trimmedLogin = loginToFind.trim();
-    if (!validator.isAlphanumeric(trimmedLogin, 'en-US', { ignore: '-' })) {
-      throw new Error('Invalid login: Only alphanumeric characters and dashes are allowed');
-    }
-
-    // Optionally validate org if needed
-    const trimmedOrg = orgToMatch.trim();
-    if (!trimmedOrg) {
-      throw new Error('Invalid organization provided');
-    }
-
-    // Retrieve the Member model once instead of on every method call.
-    const Member = mongoose.model<MemberType>('Member');
-
-    try {
-      const member = await Member.findOne({ login: trimmedLogin, org: trimmedOrg })
-        .select('login name url avatar_url')
-        .exec();
-
-      if (member) {
-        return member;
-      } else {
-        throw new Error('User not found');
-      }
-    } catch (error) {
-      throw new Error('Member lookup failed with: ' + (error as Error).message);
     }
   }
 }
